Complete weighted frontier exploration + conceptual fixes
- priority cleanup removed bc it was worse
- add gamma to control tradeoff
- add voxel filter for prio map
- raycaster conceptual correction

- target search still buggy with raycasting, need to fix

+++ fuel_planner/utils/lkh_tsp_solver/include/lkh_tsp_solver/lkh_interface.h
@@ -11,7 +11,7 @@ extern "C" {
 
 namespace lkh_interface{
 int solveTSPLKH(const char* input_file);
-void readTourFromFile(std::vector<int>& indices, const std::string& file_dir);
+void readTourFromFile(std::vector<uint8_t>& indices, const std::string& file_dir);
 void writeCostMatToFile(const Eigen::MatrixXd& cost_mat, const std::string& file_dir);
 }
 #endif
